requirements.txt
langchain
langgraph
langsmith
groq
qdrant-client
streamlit
python-dotenv
requests
PyPDF2
pydantic
pytest
pytest-asyncio
httpx
langchain-groq
langchain-community
langchain-qdrant
langchain-core
langchain-huggingface
openai
openevals
IPython
